requirements.txt
watchdog-4.0.0
